Add files via upload
In this coding file, the 10-20 day moving average crossover strategy is tested. The code takes in trading information and then, using different libraries, plots the price candlesticks for the S&P500 with the 10 and 20 day moving averages. On top of this, the code adds green and red arrows whenever the moving averages intersect. The green arrow signals a buy position and the red arrow an sell or exit position. Lastly, the code incorporates a stochastic oscillator as a supporting indicator to the main strategy.